+++ cypress/support/elements/HomeElements.js
@@ -1,6 +1,10 @@
 class HomeElements {
   get title() {
-    return cy.get('*[data-testid="home-title"]')
+    return cy.get('[data-testid="home-title"]')
+  }
+
+  get mapCreationButton() {
+    return cy.get('[data-testid="home-map-creaction-button"]')
   }
 }