+++ src/main.js
@@ -1,10 +1,10 @@
+import App from './App'
+
+import Router from './router'
+import routes from './router/routes'
+
 import './styles/global.css'
 
-document.querySelector('#app').innerHTML = `
-  <style>
-    h1 {
-      animation: fade-in 1.5s linear;
-    }
-  </style>
-  <h1 data-testid="home-title">Hello, adventurer!</h1>
-`
+App.create('#app', {
+  router: (viewContainer) => Router.create(viewContainer, routes),
+})